+++ ts/packages/memory/conversation/src/index.ts
@@ -2,5 +2,6 @@
 // Licensed under the MIT License.
 
 export * from "./podcast.js";
+export * from "./podcastMessage.js";
 export * from "./importPodcast.js";
 export * from "./conversationMemory.js";
